Moving ShieldListener
